Adding 128-bit hash

+++ games/nes/castlevania/gameInstance.hpp
@@ -179,7 +179,7 @@ class GameInstance : public GameInstanceBase
   bool allowB;
 
   GameInstance(EmuInstance* emu, const nlohmann::json& config);
-  uint64_t computeHash() const override;
+  uint128_t computeHash() const override;
   void updateDerivedValues() override;
   std::vector<std::string> getPossibleMoves() const override;
   magnetSet_t getMagnetValues(const bool* rulesStatus) const;
