+++ games/nes/rc2/gameInstance.hpp
@@ -72,7 +72,7 @@ class GameInstance : public GameInstanceBase
   bool isAirborne;
 
   GameInstance(EmuInstance* emu, const nlohmann::json& config);
-  uint64_t computeHash() const override;
+  uint128_t computeHash() const override;
   void updateDerivedValues() override;
   std::vector<std::string> getPossibleMoves() const override;
   magnetSet_t getMagnetValues(const bool* rulesStatus) const;
