+++ source/play.cpp
@@ -46,7 +46,7 @@ int getKeyPress()
 
 
 void loadSolutionFile(
-                      std::map<uint64_t, uint16_t>& hashMap,
+                      std::map<uint128_t, uint16_t>& hashMap,
                       bool& hashCollisionFound,
                       uint16_t& hashCollisionStep,
                       uint16_t& hashCollisionPrev,
@@ -119,7 +119,7 @@ void loadSolutionFile(
  }
 
  // Adding current hash to the set
- uint64_t curHash = gameInstance.computeHash();
+ uint128_t curHash = gameInstance.computeHash();
  hashMap[curHash] = 0;
 
  // Iterating move list in the sequence
@@ -146,7 +146,7 @@ void loadSolutionFile(
   for (uint16_t f = 0; f < skippedFrames; f++) unpackedMoveSequence.push_back(".");
 
   // Adding current hash to the set
-  uint64_t curHash = gameInstance.computeHash();
+  uint128_t curHash = gameInstance.computeHash();
   if (hashCollisionFound == false && hashMap.contains(curHash))
   {
    hashCollisionStep = i;
@@ -244,7 +244,7 @@ int main(int argc, char *argv[])
   scrollok(stdscr, TRUE);
 
   // This storage will indicate whether a repeated hash was found
-  std::map<uint64_t, uint16_t> hashMap;
+  std::map<uint128_t, uint16_t> hashMap;
   bool hashCollisionFound;
   uint16_t hashCollisionStep;
   uint16_t hashCollisionPrev;
